+++ apps/arrivals/views.py
@@ -89,6 +89,11 @@ def toggle(request, lan_id):
     username = request.POST.get('username')
     toggle_type = request.POST.get('type')
     previous_value = request.POST.get('prev')
+    payment_type = int(toggle_type) == 0
+    arrived_type = int(toggle_type) == 1
+
+    if not payment_type and not arrived_type:
+        return HttpResponse(status=400, content=_(u'Invalid toggle type.'))
 
     lan = LAN.objects.filter(pk=lan_id).first()
     if lan is None:
@@ -96,6 +101,9 @@ def toggle(request, lan_id):
 
     if lan.is_ended():
         return HttpResponse(status=403, content=_(u'The LAN is ended, arrivals can\'t be changed.'))
+
+    if payment_type and not lan.allow_manual_payment:
+        return HttpResponse(status=403, content=_(u'The LAN does not allow manual tickets.'))
 
     user = User.objects.filter(username=username).first()
     if user is None:
@@ -105,13 +113,11 @@ def toggle(request, lan_id):
     if attendee is None:
         return HttpResponse(status=404, content=_(u'The user is not attending the LAN.'))
 
-    if int(toggle_type) == 0:
+    # Update
+    if payment_type:
         attendee.has_paid = flip_string_bool(previous_value)
-    elif int(toggle_type) == 1:
+    elif arrived_type:
         attendee.arrived = flip_string_bool(previous_value)
-    else:
-        return HttpResponse(status=400, content=_(u'Invalid toggle type.'))
-
     attendee.save()
     return HttpResponse(status=200)