+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added LAN participant statistics page.
 - Added separate option for LAN media on front page.
 - Added user info on ticket purchase page, to force users to check their info.
+- Added LAN option for allowing manual payments on arrivals page.
 
 ### Changed